+++ tests/other/LocaleTest.php
@@ -11,4 +11,16 @@
                 $this->assertTrue(defined($d), $d . ' was not defined');
             }
         }
+
+        function testEquals() {
+            $expected = ['glob_one'  => 'one',
+                         'glob_two'  => 'two',
+                         'loc_three' => 'three',
+                         'loc_four'  => 'keturi'];
+
+            $actual = PhuGlobal::$locale->getAll();
+
+            $this->assertEquals($expected, $actual, _unit_dump(['expected' => $expected,
+                                                                'actual'   => $actual]));
+        }
     }